rename reporting batch run once router
